Clean up. Put repositories in separate files
