+++ src/domain/entities/Manager.ts
@@ -1,6 +1,4 @@
 import { Id } from "./Ref";
-import { FutureData } from "../../data/api-futures";
-import { TimeRecordStatus } from "./TimeRecord";
 
 export type Manager = {
     id: Id;
@@ -8,15 +6,3 @@ export type Manager = {
     lastName: string;
     email: string;
 };
-
-export type GetManagerOptions = {
-    Id?: Id;
-    Ids?: Id[];
-    name?: string;
-    email?: string;
-};
-
-export interface ManagerRepository {
-    get(options: GetManagerOptions): FutureData<Manager[]>;
-    getById(id: Id): FutureData<Manager>;
-}
